Update code for a new situation
The application will print out an empty message if the task list is
empty.

Add a new if-block to check if the tasklist to be displayed is empty,
if it is empty the application will print out an empty list message
instead of an empty list. Empty list message will inform the user that
there is no task in his/her list at the moment.

Doing so will help to improve the user experience of the application, as
it notifies the user that he/she does not have any task currently.
Instead of displaying an empty list. which might confuse to user, as it
may seem as an error from the application.