Use assertions
To document important assumptions about files to be loaded for the application.

These are the conditions that have to be true for the
application to work properly.

Check if the files to be loaded exists and are loaded correctly,

Assertions are used because if the files fails to load or fails to
exist, it is considered to be the programmer mistake.