Add assertion to AddCommand
AddCommand has no way to verify if a Task will be added to the TaskList.

We want to ensure that a Task will be added to the TaskList instead of a
null value.

Let's add an assert statement to verify the assumption that a Task will be
added.

This is done with an assertion statement so I can fulfill the
requirement for week 5 :/

Thanks for your patience and understanding.