Add assertions to TaskList
There is no checking of valid index passed to completeTask and
deleteTask.

Although there are prechecks before the method is called,
should those checks be removed, we can prevent the program from crashing
using assert in TaskList.

Added assertions in TaskList to assert that index cannot be negative or
bigger than TaskList size.