Remove magic literals for Event and Deadline and abstract formatDate
The keywords for Event and Deadline were literals, and the code
to format the date for printing was duplicated.

This is bad for changing the keywords for Event and Deadline in
the future. It's also bad for changing the date format for printing
as someone might forget to change the format in both classes.

Let's
* Make the keywords a constant class field of the respective classes
* Abstract out the function to format the date
* Make the date format a constant of Task