Use assertions to document important assumptions
There are certain areas in the code that assume some invariants. For
example in Duke.java, the saveTasks() method assumes task can only be
of type ToDo, Deadline or Event.

Assertions help to confirm these invariants. They will indicate that
there is something fundamentally wrong with the code, should any of
them be triggered.