Refactor code and check code quality
Code is bloated and less readable in some areas. Although it works as intended,
the code can be modified to improve overall code quality.

Refactor code by following given conventions. Extract out methods and attributes and
remove logic that is not required in the code.

It is being done that way to improve readability. Here are some of the changes done:
* Extract method from hasUsedPetalBefore() called readTasks(...)
* Extract out components array from readTasks()
* Change variable isLocationPresent to isFolderAndFilePresent