Improve code quality
The code was updated in order to comply with Java Code quality.

One of the major changes was to edit some of the if-else branches where the else branch was used to specify the final option. This may lead to errors in the future.

The code has been changed such that most of else in the if-else branch means everything else and not the final option.