Remove unary constructor for Event and Deadline
The current unary constructors for Events and Deadlines use the
String.substring() method so that it can call the constructor of
their superclass.

This is pretty clunky because the same thing can be done with the
String.split() method which is a lot more understandable.

Let's
* Add methods in AddCommand to split the input strings for Deadlines
and Events
* Shift the method to check whether the keyword exists from Parser
to AddCommand
* Remove the unary constructor
* Update EventTest with the new constructor