UI related classes: update GUI to make it beutiful
The Ui classes are not well defined and the GUI looks terrible.

Let's modify the MainWindow and DialogBox classes to add more
features related to background color of panel and text field, and
change the font and color of texts.

This ensures the good looking of GUI and the texts are more
readable towards users.