Add update extension
The user cannot modify or edit the existing task in the list directly.
The only way a user can 'edit' an existing task is for him to delete
and re-input the updated task.

This extension allows the user to be able to directly
modify the existing task by making use of 'update'
command.

This is to make the application more user-friendly as
it now supports another useful feature.