Ui: add new method printInvalidTaskNumError.
IndexOutOfBoundsExceptions are common to both DeleteCommand and
DoneCommand when the user enters an invalid task number to mark as done
or delete.

Let's extract the common method body into a new method in the Ui class
named printInvalidTaskNumError to print the error message when user
enter an invalid task number.

This reduces the duplication of code and abides by the DRY principle.