Disallow adding of duplicate tasks
Duplicate tasks can be added to the task list without any warning or
exceptions raised by Duke.

Duplicated tasks may cause confusion for the user. Furthermore, if
features such as sorting (unstable) are introduced in the future, the
duplicated tasks may cause bugs that are hard to fix.

Raising an exception when the user attempts to add a duplicate task
reminds the user that the task is duplicated and prevents user errors.
Disallowing duplicated tasks also reduces the possibility of bugs in the
future.

Let's raise a user-facing DukeException when the user attempts to add a
task that has already been added, which allows the user to know that
duplicated tasks are not allowed.

Raising a DukeException is a clean way to handle the duplicated task.
Since it is user-facing, the user is also informed that such behaviour
is not allowed.

Additional information: Streams are used to check for duplicates.