Review on code quality. Revised which includes Refactoring as well.
Includes refactoring so that the code is neater and cleaner for user to read and navigate. Applied Extractions to make the functions more readable.

Remove unwanted commented codes as well.

Such that the code will not be full of comments, as reader will read more meaningful codes and comments as well.