Add error message to TaskInputWindow
There would be an exception thrown on the console before when the deadline or
event date was not selected.

Through the use of a null check, we are able to deal with the situation when
the date is not selected. It is made such that an error message will print,
but the content is not cleared.

It is done this way to prevent the case where some description has been
added and the error erases said description.

Let's
 * Error print out when date not selected
 * Clears content on submission.
 * Submission no longer closes task input window