Add assertions to methods
There was no explicit checks for many of the assumptions tha the methods made
about the parameters.

Add assertions to Task class files and to Task container files.

By adding the explicit checks, we reduce the chance of a bug slipping by undetected
as we reduce the number of incorrect assumptions we have of the code. Tasks and
their containers are selected as they will be the more frequent use case.

Let's
 * Add assertions to Task class
 * Add assertions to factory methods for Task and its subclasses