Add sort feature to Duke
It is helpful to see tasks in chronological order so that users will
know which tasks are more urgent.

Sort allows user to list the tasks by date.
All tasks that are completed are listed at the start of the list.
Sort reverse allows user to reverse the list.

Deadline and Event no longer accommodate for incorrect format of date
time inputs. An exception is now thrown and the user is informed of the
incorrect format.

Storing incorrect date time inputs as strings makes the sorting very
confusing.