Clean code and fix Ui bugs
Code has lots of methods for CLI in the Ui, though unnecessary now
with the existence of GUI.

Removing unnecessary methods can make the code simpler and more
readable for other programmers. It also makes it cleaner for debugging,
allows the code to be more expandable in the future, and reduces
unnecessary imports in other classes.