Improve code quality
Fellow developers have cited coding style violations and that
it needs to conform with the standard to ensure that codes look like
it is bye a single developer.

Final variable names are capitalised and some of the abbreviations
in the variable names are replaced by their full form.

This is so that the code conforms to the coding standards.

A new jar file is released too with all of the JavaFX add-ins.