Add BCD-Extension by adding C-Update
There is no update function to update a specific Task in the TaskList.

Considering that the TaskList cannot be sorted or ordered, the user
potentially wants to manually add Tasks in a certain order. However, if
the user wanted to change a certain Task, he would have to delete that
Task, as well as all following Tasks, to manually sort them.

Let's add an update function so that a specific Task at a given index
be changed or modified without having to delete and then add a new Task.

Incorporating an update function would result in increased flexibility
in the functionality of the application and how it can be used.