Improve code quality through refactoring
The code base has issues like long methods, excessive nesting
and long switch cases.

A refactored codebase will be better for expandability and bug
testing because it is more readable and it is easy to find out
where the problematic code lies.

Let's,
* extract the bodies of switch cases into their own private methods
so that each case can be one or two lines long
* make use of guard clauses to make the main path more prominent
* practise using single level of abstraction
* use named constants to eliminate magic numbers
* use logical names to communicate ideas better

We follow the guidelines listed here to improve readability:
https://nus-cs2103-ay2122s1.github.io/website/se-book-adapted/chapters/codeQuality.html