Update methods
The Ui class methods were originally in-charge of printing to the screen duke's responses.
However, since the introduction of the GUI version, the Ui class methods have all been converted to returning Strings instead of printing to the screen.
Now, the print functionality is added back into the methods in order to allow the CLI version of duke to continue functioning.