Refactor code class Storage to structure clearer
The parsing of the save file was done inside storage class with a lot of code duplication.

An method to do the parsing would make the code more readable.

Adding a new method in the Parser class called parseSaveFileLine would abstract out this logic from the Storage class.

Let's move the logic from Storage into this new method in the Parser class