Remove duplicate code
Originally, the readInput method was meant for the CLI version of Duke, while the getDukeResponse method was meant for the GUI version.
However, since both methods are very similar, getDukeResponse has now been updated to achieve the same outputs for the CLI version as well. Hence there is no need for readInput anymore.