Add archive function
Code does not allow you to archive current tasks and subsequently, cannot start
again with a clean slate without losing all previous progress.

Add archive functionality by using the DossiersCommand, ArchiveCommand class. Add
relevant methods to TaskList and storage class.

It is being done this way in accordance with the existing OOP format, and it allows
for more flexibility to change in the future.

The archived tasks cannot be merged with the current tasks, for that functionality
the user still has to open the text file with the archived tasks and manually shift them to the
text file of current tasks.