Add ability to update tasks
To update any task in the task list, the task has to be deleted
and a new task created. Not only is it tedious, the task's
position in the list is also not preserved.

Let's add a new update command to allow the user to update
existing tasks without having to delete them. The user will
be able to change the description, time or both of a task.

With this approach, the task's position in the list will
not be affected.