Replace Janet and Eleanor profile images with PNGs
The JPG versions of Janet and Eleanor do not have transparent
backgrounds.

This means that the chat messages displayed are less visually
appealing.

Let's replace them with PNGs in which their faces are cropped in
a circle and the rest of the image is transparent.

Using circular images is associated with chat boxes and lets the
user feel like they are really chatting with a real person.