Implement assertions
Assumptions about the program state are not verified.

Missing out on possible bugs in the code due to incorrect invariants
can result in unsafe execution, where further code works with wrong
values and can potentially cause damage.

Let's define assertions to document and check important assumptions
that should hold at various points in the code.

Using assertions is preferable over exceptions in this situation since
we are checking for mistakes made by the programmer, not the
user or environment.

The build.gradle file has also been modified to fix the problem of
Gradle not being able to launch the GUI application.