Add Edit feature
When user want to change certain information about a task, they have to delete and re-add the task with the updated information, which is quite inconvenient.

This edit feature enables user to edit the time or description of any task.

This enables user to quickly change the description or the time of their task without having to delete and re-add the task.