Storage class: add assertions in method load. Command and its subclasses: add class fields to define various error messages.
The method load in the Storage class has some assumptions made (i.e. each task is split
using | with the necessary descriptions and date and will not reach
the default case).

There are some common error messages in the subclasses for Command,
resulting in code duplications.

Let's add assertions to define the assumptions about the program state
so that the run time can verify them and minimise potential bugs.

Let's add class fields in the parent class Command or relevant subclass
to refactor the code for clearer and more readable code.