Improve code quality in the Parser class
The code in the Parser class contains too many magic literal.
The parse method is too lengthy and hard to understand.

Replacing magic literals with constant variable and abstracting
code into seperate methods make the code easier to read and maintain.

Let's replace the input commands in the Parser class with constant
variables and abstract the conditional statement to a method.

This allows us to group similar commands into a single method call for
verification.