Use Assertions
Need to ensure that certains assumption during the execution of the code is held  to be true.
The TaskList assumes that the input passed to it is not blank and this must be held true to
ensure that the correct method is called with the correct input.

Let's use assert statement in the TaskList methods that is called by the Parser handleInput
method to check for the correctness of the input that is passed to it.

It is best for the individual method to check for the correct input as the input requirement is
not the same between them.