Add assertions to Duke.java
Assertions are used to define assumptions about the program state so
that the runtime can verify them. An assertion failure indicates a
possible bug in the code because the code has resulted in a program
state that violates an assumption about how the code should behave.

If at runtime, an assertion failure is detected, the program terminates
with and error message as a potential bug is detected. And sooner
it is detected, the safer it is.