Improve code quality
Numerous portion of the code is of different level of abstraction and to streamline the various
methods to use a single level of abstraction as much as possible.

Let's extract out various method from code which are too long.

This is to simplify the varioous code and to be more friendly to read.