Add assertions
Add assertions to various points in the code

There are several assumptions that should hold at various points in the
code. These assumptions are critical for the program to run smoothly.

Adding assertions can detect possible bugs in the code because when an
assertion error is thrown, it indicates that the code has violated an
assumption about how the code should behave.

This commit also contains a bug fix for when the program loads a task
that was marked done from the save file, the task is added into the
list without being marked done.