Provide a way to undo and redo a command
Better greater experience for the user to be able undo a mistake.

Let's implement undo and redo by keeping track of previous states that we can revert
the TaskList back to.

By keeping track of all the previous state, this will allow the user to undo/redo multiple
commands back or forth.