Add java assert statements
Java assert statements help to document important assumptions
that should hold at various points in the code. This leads to
less error prone code.

Let's add assertions to the classes that handle the Duke tasklist
logic.

The assertions are to be done in these classes as that is where the
main application logic resides.