Add assertions into program to ensure the state of program is correct during runtime
The program state may be different at times due to the randomized nature of user interaction
with the program. This may cause certain corner cases to slip out of the error handling in place.

Assertions are added at locations which are most likely to encounter errors due to user input.

This is done to optimize the code quality and ensure that assert statements do not clutter
up the code at areas in which there is no inherent need for a state check.