Improved the code quality as per the styling convention
When being reviewed by other people, the code should be of a conventional standard
so as to avoid any erroneous interpretation of code. This commit also ensures that
the code stays up to date with the latest style guide to ensure that it is as readable
as possible.

Inline variables, optimized imports, as well as remove redundant and unused
parameters / variables.

This is done as per the styling guidelines