Add Assertions
Right now, we do not check if our Task Type is not null in our HelpBotInvalidTaskTypeException. This could lead to users being confused if I make a mistake of creating a task without the right type.

To fix this, we assert `type != null` in the constructor to make sure that this forces the application to exit instead of printing the wrong error.

This prevents confusion in the case of my error.