Code quality: structure code logically, add more guiding comments, add empty line between description and parameter section in javadoc comments
The body of most methods is written as one block with no line breaks
in between to identify the various parts of the body that handle
different parts of the logic. Comments are also minimal, which makes it
hard to easily understand the logic of the code.There is also no empty
line between the description and parameter section in javadoc comments.

Grouping method bodies according to the logic they handle and adding
line breaks to separate the parts improves readability of the code.
Adding more guiding comments and empty lines between the description and
parameter section of the javadoc comments also improve readability.

Let's,
* group method bodies according to the logic they handle and add line
breaks to separate the parts
* add more in-line comments for non-trivial lines of code
* add empty lines between the description and parameter section of the
javadoc comments