Add command: add exception for unknown task type
The getTaskType identifies the added task as a Deadline if it is not
instance of an Event or Todo.

However, it does not handle the case if the task is also not a Deadline.

Adding an exception at the point of identifying an unknown task type
will catch the unexpected behaviour early, allowing for better debugging
in case of future regressions.

Let's add and throw an exception upon identification of an unknown
TaskType in the getTaskType method of AddCommand.