Fix formatting in comments
Javadocs comments are inconsistently formatted. Some have missing periods, missing capitalization, missing linebreaks, or poor grammar.

Standardising the format of these comments improves the code quality.

As a step towards such standardisation, let's ensure that all lines in javadocs comments start with a capital letter, and end with a period. There should also be a line break between the description and the first parameter. Furthermore, the first word of the description should be an active word. Doing so will provide a standard to write future javadocs comments.