Code quality: structure code logically and store string outputs as constants
The body of most methods is written as one block with no line breaks
in between to identify the various parts of the body that handle
different parts of the logic. Specific strings at different parts of the
code also makes it harder to understand why the string is returned
there, as well as to modify it.

Grouping method bodies according to the logic they handle and adding
line breaks to separate the parts improves readability of the code.
Storing the strings to be returned as aptly-named constants and then
calling on the constants themselves makes it easier to understand and
modify the strings.

Let's,
* group method bodies according to the logic they handle and add line
breaks to separate the parts
* add private constants outside of the methods and assign them to the
various string outputs