Code quality: structure code logically, store string outputs as constants, add empty line between description and parameter section in javadoc comments and add more guiding comments
The body of most methods is written as one block with no line breaks
in between to identify the various parts of the body that handle
different parts of the logic. Specific strings at different parts of the
code also makes it harder to understand why the string is returned
there, as well as to modify it. There is also no empty line between the
description and parameter section in javadoc comments. Comments are
also minimal, which makes it hard to easily understand the logic of the
code.

Grouping method bodies according to the logic they handle and adding
line breaks to separate the parts improves readability of the code.
Storing the strings to be returned as aptly-named constants and then
calling on the constants themselves makes it easier to understand and
modify the strings. Adding more guiding comments and empty lines between
the description and parameter section of the javadoc comments also
improve readability. Adding more guiding comments and empty lines
between the description and parameter section of the javadoc comments
also improve readability.

Let's,
* group method bodies according to the logic they handle and add line
breaks to separate the parts
* add private constants outside of the methods and assign them to the
various string outputs
* add empty lines between the description and parameter section of the
javadoc comments
* add more in-line comments for non-trivial lines of code