Add ability to detect duplicate tasks
Whenever a user adds a task, it can get confusing when there are multiple tasks
of the same name. Even if a user has multiple tasks of the same name, it is not
wise to add them as the same task as this is not efficient at all. Thus, it is mandatory
to add tasks of differing names from this commit forth.

Add a way to check the presence of duplicate tasks and disallows users from
creating a task that is already existing.

This is done so that the task list will be easy and intuitive, with clear distinctions
between the tasks