Add new exception classes that inherit from DukeException
There is only a singular DukeException class to account for all the possible exceptions specific to Duke.

However, this is very restrictive and not a good style of coding by OOP rules.

In order to improve this exception issue, three new separate exception classes that inherit from DukeException is created. One handles the case where the given input is completely unrecognised, another handles the case where the description of the task we are adding is missing and the last handles the case for inputs that require an index but do not have it.

By separating the possible exceptions into separate classes, it is now clearer exactly which exception occurred and what is exactly the problem should it arise in Duke.