Improve code quality
Code quality has not been critically examined.

It is important to ensure the written code is of a high quality.
This means the code should be readable with the appropriate comments,
and unsafe practices should be avoided.

Let's,
* Critically examine all the written code following the topics learnt
in the module, and ensure all written code adheres to the code
quality guidelines.