Utilise Java Streams in Duke
Java provides Stream functionality as a way of pipelining methods on a
sequence of objects. Methods that iterate through the task list might
be suitable for using Streams.

We thus should use the functionality provided to us by Java to improve
clarity.

Let's
* convert the logic for showList, showDateFind and showKeywordFind
methods in Ui to utilise Java Streams
* make minor changes to methods in TaskList to remove redundancies