Add assertions to multiple functions
Assertions are not being used in the code.

Using assertions will allow us to verify that each part of the code is working as intended on each change. And allow us to catch more bugs.

As a step towards this more automated bug catching, let's set up some assertions to ensure that bugs don't go unchecked.