Add assertions to some parts of the code
This is to ensure that when the code are modified later, the invariants are not violated.

Checking the tasks before and after execution of commands allows us to ensure that the
number of tasks are updated as expected. Checking the existence of the file after storage
is created allows us to confirm that a file is created.

Let's,
* check the number of tasks before and after the execution of a command
* check if a file exists after storage is initialised.