On launch: data from data.txt is not being read into Duke
Duke is able to write to file but it must also be able to read the file
for the next use

A quick change was made to the constructor of Duke so that the given
file path is read correctly when the GUI launches.

Previously, before the GUI was added, the task list was only set up to read the
data file information when it started run and not when it was created.

Therefore, this fix also makes the run() command in Duke simpler if it
is swapped back to the CLI version, although it should be removed soon.