Improve code quality of Parser class
The parse method of Parser was very long, affecting its readability.
Separating it into smaller methods and practicing the
Single Level of Abstraction Principle here will improve the readability.

Let's create new methods to be used in the parse method to make it
more readable and improve the code quality.