Add Assertions Classes do not have any assertions to check for assumptions made.
Failing to check for these invariants or assumptions made could result
in bugs that could cause the program
to crash.

Add assertions to check for and document assumptions that should hold
up at various points in the code.

Assertions help check the state of the program at a particular point in
the code and can be enabled or disabled accordingly.