Refactor UI class functions to be more basic and reusable
Most functions in the UI class handle both the string formatting of
the text and printing it.

The GUI does not need to print the text to the terminal, it just needs
the text.

I will separate out the print to Terminal functionality into a separate
function so the main functions will only return the formatted text.