Add reminder functionality
Provide a way for user to get reminders about tasks.

Add a new command 'remind' to get the tasks for a certain period.
For example, 'remind today' will return the tasks for today.

Adding a period will allow users to see all the tasks for the period
and be more focused.

The supported commands are 'remind today', 'remind tomorrow',
'remind week' and 'remind month'.