Complete A-CodeQuality
Among various dimensions of code quality, such as run-time efficiency, security, and robustness, one of the most important is understandability. High code quality means that the code can be understood with ease.

I have added an abstract method to my enum class that is responsible for returning the correct offset value. This means that "magic numbers" throughout the Parser class can be removed. I have also initialised variables containing the offset value in the Storage class for the same purpose.