Add in assertion to check for assumptions about the program state
The checks is done during runtime.

Checking for bugs during runtime is important during production. A failure indicates a possible bug in the code because the code has resulted in a program state that violates an assumption about how the code should behave.

Let's add in assertion to test for bugs during runtime.

Assertion is introduced into the necceary parts of the codes to check for those assumptions.