Refactor code to match code quality standards
Code contains violations of code quality standards, such as
use of magic literals and complicated expressions.

This can reduce readability and maintainability.

Various improvements have been made to improve code quality,
such as extracting magic literals to constants, and extracting
code to separate methods.

This is done in accordance with the resources and materials
provided in the CS2103 module with regards to code quality.