Tasklist, Storage and Parser class: Add assertions
The above-mentioned classes make unchecked assumptions such as the
taskNUmber being more than 0 and the filePath being writable in storage

Should these criterias not be checked and indeed fail, there could be serious
bugs occuring during runtime.

I will add assertions to check for the assumptions made in Tasklist,
Storage and Parser class to ensure that the program stops if these
assumptions are not met.