Improve code quality based on coding standard
Some areas of the code had improper and poor code comments, or unsafe coding
practices used.

Such low quality code can result in bugs, as well as increased difficulty
in maintaining and extending the codebase.

We should ensure that all code follows the coding standard so that it is
easier for others to work on the same codebase in the future.