Update main application class name for gradle
The main application class name for build.grade is Duke.

Duke runs the command line user interface while Launcher runs the GUI version. Since javaFx is implemented, the command line version should not be the default one to run.

Let's replace the application class name in build.gradle with Launcher.

Launcher is preferable than Duke as it provides a GUI for the app.