Add assertions to Storage, Command, and Task classes
There are a few classes that make implicit assumptions about the code
behaviour. The save() method in Storage class assumes that the list of
task given has been initialised. The constructors for the Command and
Task classes assume that the input provided is not empty.

Explicitly showing these assumptions via assertions improves the code
readability.