Improve code quality
There are several portions of code with highly nested code or overly complicated code.

Reducing the nesting and complexity of the code improves the code quality and makes it more readable.

As a step towards better code quality, let's break down complicated expressions into multiple simple expressions, and limit the amount of nesting in each block of code.