Add remind command and organised code stucture
Originally, the helper functions took in a string. However, this led to some problems when there was no nextLine in the scanner for this string to be generated.
Since it is more convenient for the helper methods to handle this exception, I have updated the scanner to become a class variable to allow the helper methods to access it.