Add assertions in Storage.java file
Many problems may occur during file loading and storing, such as file not being found, parent dir does not exist, taskList.txt format is incorrect, and so on.

Adding assertions to this part of the code allow us to better check some exceptions and ensure that the program runs normally.