Update GUI to include reminder function
Reminder command lists out reminders in a task list format as with the
other commands.

GUI is refactored to allow user to click on a button to get reminder on
upcoming tasks, with the dates sorted in ascending order. The earliest
task is at the top.

User can easily view reminders with the earliest one being at the top of the
reminder list.