Improve code quality and enhance code readability
There are a number of classes that have methods comprised of
unnecessary deep-nesting and complicated code logic. The code is
sometimes difficult to follow due to unsuitable variable names.

The task type and stored time in Task classes are displayed using
unnecessarily complicated ternary operations. This pattern is
repeated in the showTaskAdded() method of UserInterface class.

The code logic of execute() methods in TodoCommand, DeadlineCommand,
and EventCommand classes are not obvious and unnecessarily verbose.

Extracting more variables and methods to deal with these cases
significantly improves code quality and enhances code readability.