Implement BCD-Extension Requirements
Add tagging functionality to Duii, which will tag tasks with a certain word with a specified tag.
Update Storage to be able to load tasks with tags.
Update FindCommand to also include tasks with the keyword as part of its tag.
Added TagCommand which encapsulates the tagging command.

Add feature in which toDos can be specified with a duration.
Update toDo class to encapsulate duration in its class.

Add help command to Duii, which introduces the bot and lists all the available commands, as well as usage examples.
Add HelpCommand class which encapsulates the command when the help command is entered.