Use streams in DateCommand.execute() method
DateCommand.execute() method find the tasks on a specific day by scanning through all the tasks in the list. This part of code is a bit too long and the efficiency is low.

Using streams in this method allows us to make the code clear. It also allows us to use parallel stream to speed up this method if possible.