Restructure code to improve readability and code quality.
Current code in the Command classes have many duplicate error strings which can be reduced by
putting them in the Ui class. The Command classes and TaskList class also do not follow the SLAP principle.

The Command and TaskList classes have been further abstracted such that all error messages are in the Ui class.
The Ui class now contains all the strings which Duke will return when executing a command or returning error messages.
This is so anything to do with Duke's replies to the user can be found all in one place and it is easier to fix bugs if any.
The code is also now easier to read and much cleaner.