Add extensive JUnit testing
Current testing only meets the requirements but does not allow me to carry out
extensive testing on different components of Petal for regression and unit testing
purposes.

Add more JUnit test cases. Add more test classes that test the main classes, such as
CalendarTest, TaskListTest, and TaskTest.

It was done this way to ensure that JUnit works and every class can be tested in
isolation (with stub classes).