Assertions: add relevant assert statements in methods
Assertions is needed to check for potential errors in methods.

Some assertions help locate errors where assuptions are not fulfil.

Let's,
Add some relevant assert statements in methods to test for potential
assumption mismatches. For example, check that for each type of command,
the index used to access the descriptions and time (if applicable) is
not out of bound.