Add assertions in certain commands to ensure userArgument is not null
It is possible for a developer to pass in null to the constructor of a Command
This is undesired as certain commands require that there be an either an empty
string, or a string as the argument.

Let's use assertions to ensure that future developers do not pass in null
as the userArgument to the Commands.