Add assertion for some methods.
Add assertions for TaskList to check the number of tasks in the TaskList
is non-negative. Add assertion to parser class to ensure the string to
parse is not empty. Duke class also has an assertion to check the
parsing result to be positive integer.

Assertions check the code fulfils the expectations.