Add assertions to ensure parser loaded properly.
As parse() function is a public function, it may be called without correct states being initialized.

Let's add assertions to make sure that tasklist and storage is initalized before parsing any input.

Assertions are not part of the main code but can be used to make sure certain things are running correctly.
It also runs fast and is fast to implement.