Add duplicate task detection.
Users might accidentally add tasks multiple times.

This would make things confusing and troublesome for them.

Add a feature that detects when the user adds an item that is identical to one that is in the list, and display a confirmation popup to ensure the user wishes to add another task identical to one already in the list.