Add assertions to Command classes
DeleteCommand, DoneCommand and FindCommand classes execute a database
update/query that returns a Task or List of Tasks. The exec methods
do checks on invalid arguments.

We are not entirely certain that the database updates/queries will
somehow return null by mistake.

Assertions are added to further ensure that the database updates/
queries returns a non-null value.

Assertions can prematurely terminate the process before it cascades
to situations/problems that are hard to reverse or fix,