Command class: refactor into Command enum
Command class contains String constants for the commands. This reduces the flexibility of the
chatbot to specify which inputs match the commands.

Let's change the Command class into a Command enum, and pass the responsibility of matching the
user inputs to commands to the Parser class.

Letting the Parser class to handle the matching of inputs to commands is appropriate because it
has a getCommand method which is called by the Blue class to get the Command. The level of
abstraction is also consistent within the Parser class.