Add assertions
There are some preconditions in the code that must be satisfied before the program can run successfully.

Adding assertions will help us define assumptions about the program state so that the runtime can verify them. An assertion failure indicates a possible bug and the sooner the execution stops, the safer it is.

Let's,
* add assertions to test assumptions about the program